integrations.django: http rendering monkey patch: fix linting, move template patch to templates file, update template rendering op
